requirements.txt
amqp==2.5.2
asgiref==3.2.7
attrs==19.3.0
bcrypt==3.1.7
billiard==3.6.3.0
cached-property==1.5.1
celery==4.4.2
certifi==2020.4.5.1
cffi==1.14.0
chardet==3.0.4
cryptography==2.9.2
Django==3.0.6
django-filter==2.2.0
djangorestframework==3.11.0
docopt==0.6.2
idna==2.9
importlib-metadata==1.6.0
jsonschema==3.2.0
kombu==4.6.8
Markdown==3.2.2
marshmallow==3.6.0
paramiko==2.7.1
psycopg2-binary==2.8.5
pycparser==2.20
PyNaCl==1.4.0
pyrsistent==0.16.0
pytz==2020.1
redis==3.5.2
requests==2.23.0
six==1.15.0
sqlparse==0.3.1
texttable==1.6.2
urllib3==1.25.9
vine==1.3.0
websocket-client==0.57.0
zipp==3.1.0
